Update ports to be loaded dynamically
To make this work I had to add explict dependencies to each port so
that the can be built in the correct order.  Previously we were relying
on the ports being listed in a specific order to work around this issue.

This paves the way for ports to be maintained out of tree, and for
them to be added/removed by simply adding/removing files from this
directory.